requirements.txt
streamlit
streamlit-drawable-canvas
opencv-python-headless
numpy
Pillow







